Data.Index

Description

The indices supplied by this module are static type and value-level linked lists (which is not a concern due to unapologetic inlining of the functions operating on them, which, due to their small static nature, does not cause a code explosion).

Indices look like x:.y:.Z :: i:.j:.Z . The value is the actual index used, and the type is the upper-bound on that index.

For instance, one index of a 4x4 matrix is 2:.2:.Z :: 4:.4:.Z , and another index for a 1024x1024x1024 cube is 512:.512:.512:.Z :: 1024:.1024:.1024:.Z

Core

class (Bounded n, Integral n, Show n, Read n, Ix n, Dim n) => Index n

A single class encompassing all the instances an index should have.

class Ord n => Dim n where

Methods

zero :: n

All zeros

unit :: n

All ones

rank :: n -> Int

The number of dimensions in an index

size :: Proxy n -> Int

The size of the index

reflect :: n

Reify a type-level index into a value-level one

next' :: n -> n

Same as succ, but there are no boundary checks, so when maxBound is hit, it will wrap around to minBound / zero.

prev' :: n -> n

Same as pred, but there are no boundary checks, so when minBound / zero is hit, it will wrap around to maxBound.

toIndex :: n -> Int

fromIndex :: Int -> n

correct :: n -> n

intersect :: n -> n -> n

class Rank a b where

Methods

setBound :: a -> b

Retain the rank, but change the upper bound

data a :. b

Index constructor, analogous to :

The Applicative and Monad instances multiply in their bind operations, and their 'return'/'pure' creates an index whose first dimension is 1.

Constructors

!Int :. !b

data Z

The zero index, used to end indices, just as '[]' ends a list.

Constructors

Z

Ranges

class (Dim n, Range (ToPeano (Size n))) => Ranged n

Types that support static range operations

class Range n

data Peano

Peano numbers

Constructors

Zero

Succ Peano

type family ToPeano n :: Peano

Convert a Nat to a type-level Peano

type family Size dim :: Nat

Compute the size of an index

Static

sfoldlRange :: Ranged o => Proxy o -> (b -> o -> b) -> b -> b

See foldlRange

With optimisations, this compiles to an unrolled loop

sfoldrRange :: Ranged o => Proxy o -> (o -> b -> b) -> b -> b

See foldrRange

With optimisations, this compiles to an unrolled loop

swithRange :: (Applicative m, Ranged o) => Proxy o -> (o -> m ()) -> m ()

See withRange

With optimisations, this compiles to an unrolled loop

Runtime

foldlRange :: Dim a => Proxy a -> (b -> a -> b) -> b -> b

Eager left fold over a range

 foldlRange r f z == foldl' f z (asProxyTypeOf range r)

foldrRange :: Dim a => Proxy a -> (a -> b -> b) -> b -> b

Lazy right fold over a range

 foldrRange r f z == foldr f z (asProxyTypeOf range r)

withRange :: (Applicative m, Dim a) => Proxy a -> (a -> m ()) -> m ()

Compute something from a range

Over Int indices

Static

sfoldlRangeIndices :: Ranged o => Proxy o -> (b -> Int -> b) -> b -> b

See foldlRangeIndices

With optimisations, this compiles to an unrolled loop

sfoldrRangeIndices :: Ranged o => Proxy o -> (Int -> b -> b) -> b -> b

See foldrRangeIndices

With optimisations, this compiles to an unrolled loop

swithRangeIndices :: (Applicative m, Ranged o) => Proxy o -> (Int -> m ()) -> m ()

See withRangeIndices

With optimisations, this compiles to an unrolled loop

Runtime

foldlRangeIndices :: Dim a => Proxy a -> (b -> Int -> b) -> b -> b

Strict left fold over the raw indices under a range

foldrRangeIndices :: Dim a => Proxy a -> (Int -> b -> b) -> b -> b

Lazy right fold over the raw indices under a range

withRangeIndices :: (Applicative m, Dim a) => Proxy a -> (Int -> m ()) -> m ()

Compute something using the raw indices under a range

Utility

range :: Dim n => [n]

The range of an index

 range = foldrRange Proxy (:) []

srange :: Ranged n => [n]

Statically generated range of an index

 srange = sfoldrRange Proxy (:) []

Syntax

dim :: QuasiQuoter

Expands to a Proxy with the phantom type being the dimension specified Works in types and expressions.

Examples:

 id [dim|3 4 5|] ==> id (Proxy :: Proxy (3:.4:.5:.Z))

 Proxy :: [dim|3 4 5|] ==> Proxy :: Proxy (3:.4:.5:.Z)

module Data.Proxy